Update credentials type
With the upcoming release of the SDK, we no longer have
SelfAttestedClaims and AttestedClaims. It's a single Credential type.

This commit migrates to this version.
