+++ package.json
@@ -6,7 +6,7 @@
     "@metamask/detect-provider": "^1.2.0",
     "@polkadot/util": "^6.1.1",
     "@polkadot/util-crypto": "^6.1.1",
-    "@trustfractal/sdk": "^0.4.2",
+    "@trustfractal/sdk": "^0.5.0",
     "bfj": "^7.0.2",
     "camelcase": "^6.1.0",
     "css-loader": "4.3.0",
