+++ scripts/build.js
@@ -42,7 +42,6 @@ const isInteractive = process.stdout.isTTY;
 if (
   !checkRequiredFiles([
     paths.appPopupHtml,
-    paths.appUploadHtml,
     paths.manifestJson,
     paths.appIndexJs,
     paths.appBackgroundJs,
