+++ scripts/start.js
@@ -37,7 +37,6 @@ const isInteractive = process.stdout.isTTY;
 if (
   !checkRequiredFiles([
     paths.appPopupHtml,
-    paths.appUploadHtml,
     paths.manifestJson,
     paths.appIndexJs,
     paths.appBackgroundJs,
