+++ scripts/utils/copyPublicFolder.js
@@ -4,8 +4,7 @@ const paths = require("../../config/paths");
 function copyPublicFolder(buildFolder) {
   fs.copySync(paths.appPublic, buildFolder, {
     dereference: true,
-    filter: (file) =>
-      file !== paths.appPopupHtml && file !== paths.appUploadHtml,
+    filter: (file) => file !== paths.appPopupHtml,
   });
 }
