+++ src/index.ts
@@ -9,7 +9,6 @@ if (!environment.IS_DEV) {
 }
 
 require("./popup");
-require("./upload");
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
