+++ src/models/Connection/types.ts
@@ -1,6 +1,5 @@
 enum ConnectionTypes {
   GET_VERIFICATION_REQUEST_BACKGROUND = "GET_VERIFICATION_REQUEST_BACKGROUND",
-  HAS_CREDENTIAL_BACKGROUND = "HAS_CREDENTIAL_BACKGROUND",
   SETUP_PLUGIN_BACKGROUND = "SETUP_PLUGIN_BACKGROUND",
   VERIFY_CONNECTION_BACKGROUND = "VERIFY_CONNECTION_BACKGROUND",
   GET_BACKEND_SESSIONS_INPAGE = "GET_BACKEND_SESSIONS_INPAGE",
