--- src/models/Credential/status.ts
@@ -1,7 +1,0 @@
-enum CredentialsStatus {
-  VALID = "VALID",
-  PENDING = "PENDING",
-  INVALID = "INVALID",
-}
-
-export default CredentialsStatus;
