--- src/models/Credential/versions.ts
@@ -1,6 +1,0 @@
-enum CredentialsVersions {
-  VERSION_ONE = "V1",
-  VERSION_TWO = "V2",
-}
-
-export default CredentialsVersions;
