+++ src/popup/components/Credentials/index.tsx
@@ -4,8 +4,6 @@ import Credential from "@popup/components/common/Credential";
 import History from "@popup/components/common/History";
 import TopComponent from "@popup/components/common/TopComponent";
 
-import CredentialsVersions from "@models/Credential/versions";
-
 import { ICredential, IRequest } from "@pluginTypes/index";
 
 const RootContainer = styled.div`
@@ -20,26 +18,12 @@ export type CredentialsProps = {
 function Credentials(props: CredentialsProps) {
   const { credentials, requests } = props;
 
-  const sortedCredentials = credentials.sort(
-    (credentialA: ICredential, credentialB: ICredential) => {
-      if (credentialA.version === credentialB.version) {
-        return 0;
-      }
-
-      if (credentialA.version === CredentialsVersions.VERSION_ONE) {
-        return 1;
-      }
-
-      return -1;
-    },
-  );
-
   const getCredentialRequests = (id: string) =>
     requests.filter((request) => request.request.credential!.id === id);
 
   return (
     <TopComponent>
-      {sortedCredentials.map((credential: ICredential) => {
+      {credentials.map((credential: ICredential) => {
         const credentialsRequests = getCredentialRequests(credential.id);
 
         return (
