+++ src/redux/stores/application/reducers/app/selectors.js
@@ -10,31 +10,6 @@ export const isSetup = createSelector(
   (app) => app.setup,
 );
 
-export const getTokensContractsAddresses = createSelector(
-  (state) => state.app,
-  (app) => app.addresses.erc20,
-);
-
-export const getStakingContractsAddresses = createSelector(
-  (state) => state.app,
-  (app) => app.addresses.staking,
-);
-
-export const getClaimsRegistryContractAddress = createSelector(
-  (state) => state.app,
-  (app) => app.addresses.claimsRegistry,
-);
-
-export const getAttesterAdddress = createSelector(
-  (state) => state.app,
-  (app) => app.addresses.issuerAddress,
-);
-
-export const isStakingEnabled = createSelector(
-  (state) => state.app,
-  (app) => app.addresses.staking.enabled,
-);
-
 export const getVersion = createSelector(
   (state) => state.app,
   (app) => app.version,
