+++ src/types/plugin/Credential.ts
@@ -1,27 +1,8 @@
 import { ISerializable } from "./Common";
 
-import {
-  IAttestedClaim as ISDKAttestedClaim,
-  ISelfAttestedClaim as ISDKSelfAttestedClaim,
-} from "@trustfractal/sdk";
-import { ITransactionDetails } from "./Transaction";
+import { ICredential as ICredentialSDK } from "@trustfractal/sdk";
 
-export interface ISelfAttestedClaim
-  extends ISerializable,
-    ISDKSelfAttestedClaim {
+export interface ICredential extends ISerializable, ICredentialSDK {
   id: string;
   level: string;
-  version: string;
-  revoked: boolean;
 }
-
-export interface IAttestedClaim extends ISerializable, ISDKAttestedClaim {
-  id: string;
-  level: string;
-  version: string;
-  status: string;
-  transaction?: ITransactionDetails;
-  valid: boolean;
-}
-
-export type ICredential = ISelfAttestedClaim | IAttestedClaim;
